﻿Objective-c wrappers for GPIO control
An instance of CHGPIO is intended to control the operation of a single GPIO on the raspberry pi.  The main decision to be made was the choice of GPIO library.  I needed to work in C and tried both pigpio (http:/abyz.co.uk/rpi/pigpio/) and the bcm2835 library (http://www.airspayce.com/mikem/bcm2835/).

Pigpio
I had a lot of trouble getting the pigpio libraries to link.  Reading to try and solve that led me to a note that the c interface is dependant upon the pigpiod running as root.  That seemed a bit combersome so I moved on.

BCM2835 library
This linked without fuss and and it was quick to get a simple objective-c wrapper working.  This is the choice for now.

The only other decision was whether to stick with GPIO numbering, or use the pin numbers of the raspberry pi.  I stuck with GPIO numbering because it's more straight forward for me.  The bcm2835 library has a set of marcos to convert between GPIO numbers and Pi pin numbers if needed.