New setting - schedulePath
The schedulePath settings will tell CyberHose where to load the irrigation schedule from.

The "schedulePath" name is scanned for in the settings file and then the "=" and then the value of the variable is scanned from between the single quotes.  This method was used because means that the variable can be on anyline in the settings file.  It also enforced some level of formatting in the file.

Enforcing the formattinghas been done using muptiple scans and checks.  This seems cumbersome and may be made more efficient in the future.  Alternatively the future may show the enforcement of variable formatting to be uneccessary, especially if not other settings are required.  If further variables are added the algorithm for retrieving them should made into a separate method.