New class - CHScheduleLine
The class will store each of the variables from a single line of the CyberHose schedule.  They are decoded from the file and stored in named variables for easy getting.  The frequency variable is stored as hours, and the duration as minutes.  This is to put decoding in the startup phase and make the main daemon code simpler.