CHScheduleLine decode 'freq', 'dur', and 'days'
Days is now stored as an array of NSString objects representing the days to water.  This was done because NSArray includes methods to find objects easily, so testing a series of day variables won't be necessary.  At this time it seems that scheduling for days will be complex no matter how they are coded, thus I have made decoding and searching simple.  This may change in the future if scheduling computations can be simplified by accepting a more complex decode or search.