Added 'time' to the schedule file
If 'days' is going to be a usefule variable a start time is needed.  The 'time' variable has been added to the schedule file.  It will hold the start times in 4 digit time, separated by a space.  For example:

line      | freq | dur | days | time      | s0(tank) | s1(mains) | s2(grey)
0(fruit)  | 2w   | 1h  | -    | 1900      | 1        | 1         | 1
1(bonsai) | 1d   | 10m | -    | 1830      | 1        | 1         | 0
2(veges)  | -    | 30m | 84   | 1900 0700 | 1        | 1         | 1

The time entry for line 2(veges) will irrigate at 0700 and 1900 on the selected days.  Where days are not specified (i.e. using frequency) the first 'time' value will be used as for the inital irrigation after CyberHose starts up.