New class - CHTime
This class was added to simplify the management of times within the schedule.  NSDate was deemed as uneccessarily complex because a full date is not expected to be needed.  Instances of CHTime store a non-mutable 4 digit time and will provide/get any required representations of that time.  Currently only initialisation of CHTime is implmented, getters will be implemented alongside the timing features of CyberHose.