Removed use of NSTimer
Instances of NSTimer seem to require an NSApplication runloop.  Including an NSApplication adds a lot of overhead (i.e. AppKit) and testing it showed that it is not really appropraite because it seems to enforce creating GUI applications (even when no GUI objects are created), and blocks access to syslog.

In contrast, using sleep() withing a while(1) polling loop seems to consume little CPU resources.

Thus the plan going forward is to use the while(1) loop in this commit and write methods to timing checking timing against the system clock.  Event driven timing might still be possible, but given the poll loop has to be there to hold the process that probably the best way to go.