New class NSEvent
The cyberhose daemon will keep a list of timing events that include triggerTimes and actions.  That list will be polled by the main loop.  For this to work CHSchedule needed methods to generate trigger times as NSDate objects.

Before entering the polling loop cyberhose will create NSEvents for each CHScheduleLine in the CHSchedule and put them in an array.  The polling loop will poll that array.  Each time a start event triggers and stop event will be created for that scheduleLine, and each time a stopEvent is triggered the next start event will be created and added to the list.

A setting has been added to allow the user to specify how many lines can be irrigated simulataenouly so that an approriate water pressure can be maintined if cyberhose is controlling many irrigation lines.  If the maximum number of lines is alreadying being irrigated when the start event triggers, it will be added to a queue.  Each time a stop event is triggered the queue will be checked for wating start events.